parsec_cuda_kernel_schedule() renamed as parsec_cuda_kernel_enqueue(). parsec_cuda_kernel_dequeue() schedules a migrated task to the gpu device.
This function will be called in __parsec_context_wait() just before parsec_current_scheduler->module.select().
This will ensure that the migrated tasks will get priority over new tasks. When a compute thread calls this function,
 it is forced to try to be a manager of the device. If the device already has a manager, the compute thread passes
the control of the task to the manager. If not the compute thread will become the manager.