Hash table: count used buckets only when needed
The used-bucket count is used to remove old tables from the search
path once they are empty. However, keeping the count for the current
head requires an atomic operation for each insert and removal for a
value that is otherwise unused while the table is the current head.
Instead, count the number of used buckets during resize so that we
can remove the table once it is empty.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>