First level migration added.
We use the exiting stage_in functionality of a task to do device to
device data trasfer. The readers, coherency_state and data_transfer_status,
of the data to be migrated, is manipulated to perform the D2D transfer.

In parsec_gpu_data_stage_in(), conditions are added to adapt it for D2D transfer
for task migration.