[RFC] profiling tools: more efficient cross-stream event matching (#372)
* profiling tools: remove unused `start' arg in matching functions

Relevant functions:
  * dbp_iterator_move_to_matching_event
  * dbp_iterator_find_matching_event_all_threads

These two functions are only ever called with `start' == 0.
Matching an end event with a start event doesn't work right anyway
and is not supported. Remove this functionality.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: remove duplicated check for profile buffer size

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: pass dbp_file_t* to refer_events_buffer, not fd

This change was made in preparation for more easily supporting a
mmap-once implementation, but this has been left for later.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: factor out dbp_iterator_set_offset, dbp_iterator_next_buffer

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: factor out dbp_events_match

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: misc minor changes and cleanup

Due to structure alignment, it's better to put members with greater
alignment restrictions at the beginning of a structure and group smaller
members at the end. In practice though, this particular change probably
does very little.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>

* profiling tools: more efficient cross-stream event matching

Performance when matching events that start in one profiling stream and
end in a different one is abysmal: O(m * n) in the worst case for m such
events and n total events. This has largely not been a problem, since
the only time this occurs is when using GPUs, and there are relatively
few such events.

In trying to profile my ongoing work with LCI, this is no longer the
case. Communication events almost always start in the progress thread
and end in the communication thread, or vice-versa. This results in many
such events, so the current behavior is infeasible.

This code relies on an observation by George Bosilca that most start
events are immediately followed by their end events. This means it is
possible to construct a list of end events that are possible matches for
cross-stream events. If the matching event for a start event is not
found in its own stream, this list can be searched to determine where to
start the search in other streams. Since the list is naturally sorted by
timestamp, a binary search can be used. I rely on bsearch and a slight
hack that stores the last array element in the search key to implement a
"fuzzy" timestamp search on a per-event-key basis. This list is
constructed lazily when it is needed and only once per stream.

Constructing the cache is O(n).
Search performance should be O(m * log m) for most cases.

There could be some edge cases where the search finds a potential match
in a stream, but the true match is on a different stream; in this case,
the current code still traverses all possibly-matching events in the
mistaken stream. There isn't much to be done about this, as we cannot
know that the matching event is not in the stream until it has been
traversed in its entirety.

Signed-off-by: Omri Mor <omrimor2@illinois.edu>